pyproject.toml
[tool.poetry]
name = "enhanced-ocr-tool"
version = "0.1.0"
description = "Enhanced OCR tool with translation and summarization features"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyQt6 = "^6.6.0"
Pillow = "^10.0.0"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
pyperclip = "^1.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
